Added optimisation for lsoa processing

+++ scripts/T3_calculation.py
@@ -51,9 +51,16 @@ def check_tree_vom_pair(chm_path: str|Path, trees_dir: str|Path) -> bool:
 def process_vom_tiles(trees_path_lst: list, tree_area: int=10, tree_height: int=3) -> gpd.GeoDataFrame:
 
     logging.info(f"Reading {len(trees_path_lst)} VOM tiles")
+    if len(trees_path_lst) > 1:
+        
+        trees_gdf_lst = [gpd.read_file(tree_path) for tree_path in trees_path_lst]
+        merged_trees_gdf = gpd.GeoDataFrame(pd.concat(trees_gdf_lst, ignore_index=True))
+    elif len(trees_path_lst) == 1: 
+        merged_trees_gdf = gpd.read_file(trees_path_lst[0])
 
-    trees_gdf_lst = [gpd.read_file(tree_path) for tree_path in trees_path_lst]
-    merged_trees_gdf = gpd.GeoDataFrame(pd.concat(trees_gdf_lst, ignore_index=True))
+    else:
+        return None
+
     geo_trees_gdf = merged_trees_gdf[(merged_trees_gdf.area > tree_area) & (merged_trees_gdf.height > tree_height)].reset_index(drop=True)
     geo_trees_gdf['treeID'] = range(len(geo_trees_gdf))
 
@@ -102,7 +109,33 @@ def count_trees(geo_level: str, geo_code: str) -> pd.DataFrame:
 
     return trees_within_buffer_df
 
+# def get_vom_trees_paths(geo_code: str, vom_tree_pair_dict: dict) -> list:
+
+#     trees_path_lst = [tree_path for _, tree_path in vom_tree_pair_dict[geo_code] if tree_path is not None]
+
+#     return trees_path_lst
+
+def get_overlapping_tiles(imd_lsoa_bua_buffer_gdf, os_5km_boundaries_gdf, geo_level, geo_code):
+    # Select one feature from imd_lsoa_bua_buffer_gdf
+    selected_feature = imd_lsoa_bua_buffer_gdf[imd_lsoa_bua_buffer_gdf[geo_level] == geo_code]
+
+    # Get the overlapping features
+    overlapping_tiles_lst = gpd.overlay(selected_feature, os_5km_boundaries_gdf, how='intersection')['TILE_NAME'].unique().tolist()
+
+    return overlapping_tiles_lst
+
+def get_vom_trees_paths(overlapping_tiles_lst: list, vom_tree_pair_dict: dict) -> list:
+
+    vom_trees_path_lst = [path_pair[1] for _,v in vom_tree_pair_dict.items() for path_pair in v if path_pair[1] is not None]
+    for tile_name in overlapping_tiles_lst:
+        translated_tile_name = translate_tile_name(tile_name).upper()
+        trees_path_lst = [path for path in vom_trees_path_lst if translated_tile_name in path]
+
+        return trees_path_lst
+
+
 def process_geo_code(geo_level: str, geo_code: str, vom_tree_pair_dict: dict,
+                     imd_lsoa_bua_buffer_gdf: gpd.GeoDataFrame, os_5km_boundaries_gdf: gpd.GeoDataFrame,
                      tree_area: int=10, tree_height: int=3) -> None:
 
     start_time = time.time()
@@ -110,23 +143,29 @@ def process_geo_code(geo_level: str, geo_code: str, vom_tree_pair_dict: dict,
     T3_dir = VECTOR_OUT_DIR / "3-30-300" / "T3"
     T3_dir.mkdir(parents=True, exist_ok=True)
     tree_count_path = T3_dir / f"T3_{geo_code}.csv"
+    if not tree_count_path.exists():
+        try:
+            overlapping_tiles_lst = get_overlapping_tiles(imd_lsoa_bua_buffer_gdf, os_5km_boundaries_gdf, geo_level, geo_code)
+            trees_path_lst = get_vom_trees_paths(overlapping_tiles_lst, vom_tree_pair_dict)
+            # trees_path_lst = get_vom_trees_paths(geo_code, vom_tree_pair_dict)
 
-    trees_path_lst = [tree_path for _, tree_path in vom_tree_pair_dict[geo_code] if tree_path is not None]
+            geo_trees_gdf = process_vom_tiles(trees_path_lst, tree_area=tree_area, tree_height=tree_height)
+            geo_trees_sdf = sedona.createDataFrame(geo_trees_gdf)
+            geo_trees_sdf.createOrReplaceTempView("geo_trees")
 
-    geo_trees_gdf = process_vom_tiles(trees_path_lst, tree_area=tree_area, tree_height=tree_height)
-    geo_trees_sdf = sedona.createDataFrame(geo_trees_gdf)
-    geo_trees_sdf.createOrReplaceTempView("geo_trees")
+            process_buildings(geo_level, geo_code)
 
-    process_buildings(geo_level, geo_code)
+            trees_within_buffer_df = count_trees(geo_level, geo_code)
 
-    trees_within_buffer_df = count_trees(geo_level, geo_code)
+            trees_within_buffer_df.to_csv(tree_count_path, index=False)
+            
+            logging.info(f"Saving file for {geo_code} with {len(trees_within_buffer_df)} records")
 
-    trees_within_buffer_df.to_csv(tree_count_path, index=False)
-    
-    logging.info(f"Saving file for {geo_code} with {len(trees_within_buffer_df)} records")
+            end_time = time.time()
+            logging.info(f"Processing for {geo_code} took {end_time - start_time:.2f} seconds")
 
-    end_time = time.time()
-    logging.info(f"Processing for {geo_code} took {end_time - start_time:.2f} seconds")
+        except Exception as e:
+            logging.error(f"Error processing {geo_code}: {e}")
     
 if __name__ == "__main__":
 
@@ -136,6 +175,7 @@ if __name__ == "__main__":
     parser.add_argument('--parallel', action='store_true', help='Run job in parallel')
     parser.add_argument('--n_workers', type=int, required=False, default=2, help='Number of workers')
     parser.add_argument('--log_level', type=str, required=False, default='WARNING', help='Logging level')
+    parser.add_argument('--buffer', type=int, required=False, default=100, help='Buffer size in meters')
 
     args = parser.parse_args()
 
@@ -144,6 +184,9 @@ if __name__ == "__main__":
     parallel = args.parallel
     n_workers = args.n_workers
     log_level = args.log_level
+    buffer = args.buffer
+
+    project_crs = 'EPSG:27700'
 
     # IN paths
     vom_dir = RASTER_IN_DIR / "Defra" / "VOM"
@@ -152,19 +195,22 @@ if __name__ == "__main__":
     T3_30_300_DIR = VECTOR_OUT_DIR / "3-30-300"
     T3_dir = T3_30_300_DIR / "T3"
     trees_dir = T3_30_300_DIR / "VOM_Trees"
+    os_5km_boundaries_path = VECTOR_IN_DIR / "OS" / "National_Grid" / "5km_grid_region.shp"
     imd_lsoa_bua_boundaries_path = VECTOR_OUT_DIR / "IMD" / "English_IMD_2019_BUA_filtered_boundaries.geojson"
     buildings_path = VECTOR_IN_DIR / "EDINA" / "Buildings_6183" / "Buildings_6183.parquet"
 
-    log_path = Path("logs/T3_calculation.log")
+    log_path = Path("logs/T3_calculation_sequential.log")
     setup_logger(log_path=log_path, log_level=log_level)
     logging.info("Calculating the 3 metric for all geographies")
     logging.debug("Reading files")
 
     chm_lad_tiles_dict = json.load(open(chm_lad_tiles_path))
-
+    os_5km_boundaries_gdf = gpd.read_file(os_5km_boundaries_path).to_crs(project_crs)
     imd_lsoa_bua_gdf = gpd.read_file(imd_lsoa_bua_boundaries_path)
+    imd_lsoa_bua_buffer_gdf = imd_lsoa_bua_gdf.copy()
+    imd_lsoa_bua_buffer_gdf['geometry'] = imd_lsoa_bua_buffer_gdf['geometry'].buffer(buffer)
     geo_level_codes = imd_lsoa_bua_gdf[geo_level].unique()
-
+    geo_level_codes = geo_level_codes[::-1]
     logging.debug("Setting up Apache Sedona")
     os.environ["JAVA_HOME"] = JAVA_HOME
     sedona = get_spark()
@@ -180,7 +226,7 @@ if __name__ == "__main__":
         logging.debug("Running in parallel")
 
         with concurrent.futures.ThreadPoolExecutor(max_workers=n_workers) as executor:
-            futures = [executor.submit(process_geo_code, geo_level, geo_code, vom_tree_pair_dict) for geo_code in geo_level_codes]
+            futures = [executor.submit(process_geo_code, geo_level, geo_code, vom_tree_pair_dict, imd_lsoa_bua_buffer_gdf, os_5km_boundaries_gdf) for geo_code in geo_level_codes]
             
             for future in tqdm(concurrent.futures.as_completed(futures), total=len(futures), desc="Regions Processed"):
                 try:
@@ -192,4 +238,4 @@ if __name__ == "__main__":
         logging.debug("Running sequentially")
 
         for geo_code in tqdm(geo_level_codes, desc='Regions Processed'):   
-            process_geo_code(geo_level, geo_code, vom_tree_pair_dict)
+            process_geo_code(geo_level, geo_code, vom_tree_pair_dict, imd_lsoa_bua_buffer_gdf, os_5km_boundaries_gdf)
